update elementwise op
