+++ cinn/frontend/decomposer/activation.cc
@@ -49,18 +49,42 @@ void relu_grad(const Instruction& instr, const DecomposerContext& context) {
   context.MapOutToOrigin(res, dx);
 }
 
+void gelu(const Instruction& instr, const DecomposerContext& context) {
+  CHECK_EQ(instr->inputs.size(), 1UL) << " 1 input tensor for " << instr->op_type;
+  CHECK_EQ(instr->outputs.size(), 1UL) << "1 output tensor for " << instr->op_type;
+  auto x        = instr->inputs[0];
+  auto output   = instr->outputs[0];
+  auto* builder = context.builder();
+
+  // x * (0.5 + 0.5 * erf(sqrtf(0.5) * x))
+  auto p_5 = builder->FillConstant<float>(x->shape, 0.5f, common::UniqName("p_5"));
+  auto p_7 = builder->FillConstant<float>(x->shape, std::sqrt(0.5), common::UniqName("p_7"));
+  auto erf = builder->Erf(builder->Multiply(x, p_7));
+  auto cdf = builder->Add(p_5, builder->Multiply(p_5, erf));
+  auto out = builder->Multiply(x, cdf);
+
+  // map the the output of decomposed operator to the original.
+  context.MapOutToOrigin(out, output);
+}
+
 }  // namespace decomposer
 }  // namespace frontend
 }  // namespace cinn
 
-CINN_REGISTER_HELPER(activation_decomposers) {
+CINN_REGISTER_HELPER(relu_decomposers) {
   CINN_DECOMPOSER_REGISTER(relu, cinn::frontend::decomposer::relu);
 
   return true;
 }
 
-CINN_REGISTER_HELPER(activation_grad_decomposers) {
+CINN_REGISTER_HELPER(relu_grad_decomposers) {
   CINN_DECOMPOSER_REGISTER(relu_grad, cinn::frontend::decomposer::relu_grad);
+
+  return true;
+}
+
+CINN_REGISTER_HELPER(gelu_decomposers) {
+  CINN_DECOMPOSER_REGISTER(gelu, cinn::frontend::decomposer::gelu);
 
   return true;
 }
