+++ cinn/hlir/op/contrib/gather_nd.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "cinn/hlir/op/contrib/gather.h"
+#include "cinn/hlir/op/contrib/gather_nd.h"
 
 #include <gflags/gflags.h>
 
@@ -47,25 +47,6 @@ namespace op {
 using common::CINNValue;
 using common::CINNValuePack;
 
-ir::Tensor Gather(const ir::Tensor &A, const ir::Tensor &B, const int &axis, const std::string &name) {
-  CHECK_EQ(A->shape.size(), B->shape.size());
-  auto res = Compute(
-      B->shape,
-      [=](const std::vector<Expr> &indices) {
-        std::vector<Expr> A_indices;
-        for (int i = 0; i < axis; ++i) {
-          A_indices.push_back(indices[i]);
-        }
-        A_indices.push_back(ir::Cast::Make(common::I32(), B(indices)));
-        for (size_t i = axis + 1; i < A->shape.size(); ++i) {
-          A_indices.push_back(indices[i]);
-        }
-        return lang::Identity(A(A_indices));
-      },
-      name);
-  return res;
-}
-
 ir::Tensor GatherNd(const ir::Tensor &A, const ir::Tensor &B, const std::vector<int> &axes, const std::string &name) {
   std::vector<Expr> out_shape = B->shape;
   out_shape.pop_back();
@@ -83,83 +64,6 @@ ir::Tensor GatherNd(const ir::Tensor &A, const ir::Tensor &B, const std::vector<
       },
       name);
   return res;
-}
-
-std::shared_ptr<framework::OpStrategy> StrategyForGather(const framework::NodeAttr &attrs,
-                                                         const std::vector<ir::Tensor> &inputs,
-                                                         const std::vector<Type> &out_type,
-                                                         const std::vector<std::vector<int>> &output_shapes,
-                                                         const Target &target) {
-  auto attr_store = attrs.attr_store;
-  CHECK(attr_store.count("axis")) << "find no attr of axis";
-  int axis = absl::get<int>(attr_store.at("axis"));
-  std::string op_name("gather");
-
-  framework::CINNCompute gather_compute([=](lang::Args args, lang::RetValue *ret) {
-    CHECK(!args.empty()) << "The input arguments of " << op_name << " compute is empty! Please check.\n";
-    CINNValuePack pack_args = args[0];
-    CHECK_GE(pack_args.size(), 2U) << "2 input tensors for " << op_name << " compute\n";
-    Expr A = pack_args[0];
-    Expr B = pack_args[1];
-    CHECK(A.as_tensor());
-    CHECK(B.as_tensor());
-    CHECK(!output_shapes.empty());
-    auto tensor_A = A.as_tensor_ref();
-    auto tensor_B = B.as_tensor_ref();
-    auto stages   = CreateStages({tensor_A, tensor_B});
-    VLOG(3) << "A shape: " << utils::Join(tensor_A->shape, ", ") << ", B shape: " << utils::Join(tensor_B->shape, ", ")
-            << ", output_shapes: " << utils::Join(output_shapes[0], ", ");
-    std::string tensor_name = UniqName("Gather_out");
-    if (FLAGS_cinn_ir_schedule) {
-      CHECK_EQ(pack_args.size(), 3U);
-      tensor_name = pack_args[2].operator std::string();
-    }
-    ir::Tensor out = Gather(tensor_A, tensor_B, axis, tensor_name);
-    std::vector<CINNValue> res;
-    stages->InsertLazily(out);
-    res.push_back(CINNValue(out));
-    CHECK(!out_type.empty()) << "Output type of " << op_name << " is empty! Please check.\n";
-    res.push_back(CINNValue(stages));
-    *ret = CINNValuePack{res};
-  });
-
-  framework::CINNSchedule gather_schedule([=](lang::Args args, lang::RetValue *ret) {
-    if (FLAGS_cinn_ir_schedule) {
-      CHECK(!args.empty()) << "The input argument of gather_schedule is empty! Please check.\n";
-      common::CINNValuePack arg_pack = args[0];
-      std::vector<Expr> vec_ast;
-      for (int i = 0; i < arg_pack.size(); i++) {
-        if (arg_pack[i].is_expr()) {
-          Expr temp = arg_pack[i];
-          vec_ast.emplace_back(temp);
-        }
-      }
-      CHECK(!vec_ast.empty());
-      ir::ModuleExpr mod_expr(vec_ast);
-      ir::IRSchedule ir_sch(mod_expr);
-      ir_sch.MergeExprs();
-      long prod_size = std::accumulate(output_shapes[0].begin(), output_shapes[0].end(), 1, std::multiplies<int>());
-      if (prod_size > 1) {
-        if (target.arch == Target::Arch::NVGPU) {
-          pe::IRCudaScheduleInjective(ir_sch, output_shapes.front(), target);
-        } else if (target.arch == Target::Arch::X86) {
-          pe::IRScheduleInjectiveCPU(ir_sch, output_shapes.front(), target, true);
-        }
-      }
-      std::vector<common::CINNValue> res{common::CINNValue(ir_sch.GetModule().GetExprs().at(0))};
-      *ret = common::CINNValuePack{res};
-    } else {
-      CHECK(!args.empty()) << "The input argument of gather_schedule is empty! Please check.\n";
-      CINNValuePack arg_pack = args[0];
-      Expr out               = arg_pack[0];
-      CHECK(out.as_tensor());
-      *ret = arg_pack;
-    }
-  });
-
-  auto strategy = std::make_shared<framework::OpStrategy>();
-  strategy->AddImpl(gather_compute, gather_schedule, "strategy.gather.x86", 1);
-  return strategy;
 }
 
 std::shared_ptr<framework::OpStrategy> StrategyForGatherNd(const framework::NodeAttr &attrs,
@@ -239,14 +143,6 @@ std::shared_ptr<framework::OpStrategy> StrategyForGatherNd(const framework::Node
   return strategy;
 }
 
-std::vector<std::vector<int>> InferShapeForGather(const std::vector<std::vector<int>> &inputs_shape,
-                                                  const framework::AttrMapType &attrs) {
-  CHECK_EQ(inputs_shape.size(), 2U) << "The input's shape size should be 2! Please check again.";
-  CHECK_EQ(inputs_shape[0].size(), inputs_shape[1].size()) << "The inputs' dims should be equal.";
-  std::vector<std::vector<int>> res{inputs_shape[1]};
-  return res;
-}
-
 std::vector<std::vector<int>> InferShapeForGatherNd(const std::vector<std::vector<int>> &inputs_shape,
                                                     const framework::AttrMapType &attrs) {
   CHECK_EQ(inputs_shape.size(), 2U) << "The input's shape size should be 2! Please check again.";
@@ -255,7 +151,7 @@ std::vector<std::vector<int>> InferShapeForGatherNd(const std::vector<std::vecto
   return res;
 }
 
-std::vector<Type> InferDtypeForGather(const std::vector<Type> &inputs_type, const framework::AttrMapType &attrs) {
+std::vector<Type> InferDtypeForGatherNd(const std::vector<Type> &inputs_type, const framework::AttrMapType &attrs) {
   CHECK(!inputs_type.empty()) << "The input's type size is 0! Please check again.";
   std::vector<Type> res{inputs_type[0]};
   return res;
@@ -265,23 +161,14 @@ std::vector<Type> InferDtypeForGather(const std::vector<Type> &inputs_type, cons
 }  // namespace hlir
 }  // namespace cinn
 
-CINN_REGISTER_HELPER(gather_ops) {
-  CINN_REGISTER_OP(gather)
-      .describe("Gather.")
-      .set_num_inputs(2)
-      .set_num_outputs(1)
-      .set_attr<cinn::hlir::framework::StrategyFunction>("CINNStrategy", cinn::hlir::op::StrategyForGather)
-      .set_attr("infershape", MakeOpFunction(cinn::hlir::op::InferShapeForGather))
-      .set_attr("inferdtype", MakeOpFunction(cinn::hlir::op::InferDtypeForGather))
-      .set_support_level(4);
-
+CINN_REGISTER_HELPER(gather_nd_ops) {
   CINN_REGISTER_OP(gather_nd)
       .describe("GatherNd.")
       .set_num_inputs(2)
       .set_num_outputs(1)
       .set_attr<cinn::hlir::framework::StrategyFunction>("CINNStrategy", cinn::hlir::op::StrategyForGatherNd)
       .set_attr("infershape", MakeOpFunction(cinn::hlir::op::InferShapeForGatherNd))
-      .set_attr("inferdtype", MakeOpFunction(cinn::hlir::op::InferDtypeForGather))
+      .set_attr("inferdtype", MakeOpFunction(cinn::hlir::op::InferDtypeForGatherNd))
       .set_support_level(4);
 
   return true;
