+++ cinn/hlir/pe/transform.cc
@@ -116,40 +116,6 @@ ir::Tensor Reshape(const ir::Tensor& A,
   return out;
 }
 
-ir::Tensor Reshape(const ir::Tensor& A, const std::vector<int>& new_shape, const std::string& name) {
-  std::vector<Expr> new_expr_shape;
-  std::vector<Expr> A_expr_shape = A->shape;
-  int input_total_size           = 1;
-  int output_total_size          = 1;
-  for (auto& i : A_expr_shape) {
-    CHECK(i.is_constant()) << "Input tensor's shape should be constant value.";
-    input_total_size *= static_cast<int>(i.get_constant());
-  }
-  for (auto& i : new_shape) {
-    output_total_size *= i;
-    new_expr_shape.push_back(Expr(i));
-  }
-  CHECK_EQ(input_total_size, output_total_size)
-      << "In op reshape, the input tensor and output tensor's total size should be equal, please check!";
-  auto res = Compute(
-      new_expr_shape,
-      [=](const std::vector<Expr>& indice) {
-        Expr offset = Expr(0);
-        for (int i = 0; i < indice.size(); i++) {
-          offset = offset * new_expr_shape[i] + indice[i];
-        }
-        std::vector<Expr> indice_a;
-        for (int i = A_expr_shape.size() - 1; i >= 0; i--) {
-          auto temp = offset % A_expr_shape[i];
-          indice_a.insert(indice_a.begin(), common::AutoSimplify(temp));
-          offset = (offset - temp) / A_expr_shape[i];
-        }
-        return A(indice_a);
-      },
-      name);
-  return res;
-}
-
 std::vector<ir::Tensor> Split(const ir::Tensor& A,
                               int axis,
                               const std::vector<std::vector<int>>& output_shapes,
@@ -981,40 +947,23 @@ ir::Tensor ScatterAdd(const ir::Tensor& input,
   return output;
 }
 
-ir::Tensor ExpandDims(const ir::Tensor& input, int axis, int num_newaxis, const std::string& output_name) {
-  int ndims = input.ndims();
-  CHECK(axis >= -ndims - 1 && axis <= ndims)
-      << "expand_dims only accept `axis` in [-x.ndim - 1, x.ndim], but got axis = " << axis
-      << ", and x.ndims = " << ndims;
-  CHECK_GE(num_newaxis, 0);
-
-  if (axis < 0) {
-    axis = ndims + axis + 1;
-  }
-  std::vector<Expr> output_shape;
-  for (size_t i = 0; i < static_cast<size_t>(axis); ++i) {
-    output_shape.push_back(input->shape[i]);
-  }
-  for (size_t i = 0; i < static_cast<size_t>(num_newaxis); ++i) {
-    output_shape.push_back(Expr(1));
-  }
-  for (size_t i = axis; i < input->shape.size(); ++i) {
-    output_shape.push_back(input->shape[i]);
-  }
-
-  return Compute(
-      output_shape,
-      [=](const std::vector<Expr>& indice) {
-        std::vector<Expr> idx;
-        for (size_t i = 0; i < static_cast<size_t>(axis); ++i) {
-          idx.push_back(indice[i]);
+ir::Tensor Gather(const ir::Tensor& input, const ir::Tensor& index, const int& axis, const std::string& output_name) {
+  CHECK_EQ(input->shape.size(), index->shape.size());
+  auto res = Compute(
+      index->shape,
+      [=](const std::vector<Expr>& indices) {
+        std::vector<Expr> A_indices;
+        for (int i = 0; i < axis; ++i) {
+          A_indices.push_back(indices[i]);
         }
-        for (size_t i = axis + num_newaxis; i < indice.size(); ++i) {
-          idx.push_back(indice[i]);
+        A_indices.push_back(ir::Cast::Make(common::I32(), index(indices)));
+        for (size_t i = axis + 1; i < input->shape.size(); ++i) {
+          A_indices.push_back(indices[i]);
         }
-        return input(idx);
+        return input(A_indices);
       },
       UniqName(output_name));
+  return res;
 }
 
 }  // namespace pe
